agent/requirements.txt
livekit-agents>=1.2.0
google-genai>=1.0.0
python-dotenv>=1.0.0
faiss-cpu>=1.7.4
numpy>=1.24.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-text-splitters>=0.0.1
pypdf>=3.17.0
unstructured>=0.10.30
aiohttp>=3.9.0